(closer to) being able to run behave tests outside project root directory
